Update plugin exports
